Use IDownstreamCompiler interface. Split out DownstreamCompilerUtil and DownstreamCompilerSet.
