+++ source/compiler-core/slang-dxc-compiler.h
@@ -1,7 +1,7 @@
 #ifndef SLANG_DXC_COMPILER_UTIL_H
 #define SLANG_DXC_COMPILER_UTIL_H
 
-#include "slang-downstream-compiler.h"
+#include "slang-downstream-compiler-util.h"
 
 #include "../core/slang-platform.h"
