+++ source/compiler-core/slang-fxc-compiler.h
@@ -1,7 +1,7 @@
 #ifndef SLANG_FXC_COMPILER_UTIL_H
 #define SLANG_FXC_COMPILER_UTIL_H
 
-#include "slang-downstream-compiler.h"
+#include "slang-downstream-compiler-util.h"
 
 #include "../core/slang-platform.h"
