+++ source/compiler-core/slang-llvm-compiler.cpp
@@ -22,6 +22,24 @@
 namespace Slang
 {
 
+class DownstreamCompilerAdapter_Dep1 : public DownstreamCompilerBase
+{
+public:
+    // IDownstreamCompiler
+    virtual SLANG_NO_THROW const Desc& SLANG_MCALL getDesc() SLANG_OVERRIDE { return m_dep->getDesc(); }
+    virtual SLANG_NO_THROW SlangResult SLANG_MCALL compile(const CompileOptions& options, RefPtr<DownstreamCompileResult>& outResult) SLANG_OVERRIDE { return m_dep->compile(options, outResult); }
+    virtual SLANG_NO_THROW SlangResult SLANG_MCALL disassemble(SlangCompileTarget sourceBlobTarget, const void* blob, size_t blobSize, ISlangBlob** out) SLANG_OVERRIDE { return m_dep->disassemble(sourceBlobTarget, blob, blobSize, out); }
+    virtual SLANG_NO_THROW bool SLANG_MCALL isFileBased() SLANG_OVERRIDE { return m_dep->isFileBased(); }
+
+    DownstreamCompilerAdapter_Dep1(DownstreamCompiler_Dep1* dep): 
+        m_dep(dep)
+    {
+    }
+
+protected:
+    RefPtr<DownstreamCompiler_Dep1> m_dep;
+};
+
 /* static */SlangResult LLVMDownstreamCompilerUtil::locateCompilers(const String& path, ISlangSharedLibraryLoader* loader, DownstreamCompilerSet* set)
 {
     ComPtr<ISlangSharedLibrary> library;
@@ -34,17 +52,19 @@ namespace Slang
         return SLANG_FAIL;
     }
 
-    typedef SlangResult(*CreateDownstreamCompilerFunc)(RefPtr<DownstreamCompiler>& out);
+    typedef SlangResult(*CreateDownstreamCompilerFunc_Dep1)(RefPtr<DownstreamCompiler_Dep1>& out);
 
-    auto fn = (CreateDownstreamCompilerFunc)library->findFuncByName("createLLVMDownstreamCompiler");
+    auto fn = (CreateDownstreamCompilerFunc_Dep1)library->findFuncByName("createLLVMDownstreamCompiler");
     if (!fn)
     {
         return SLANG_FAIL;
     }
 
-    RefPtr<DownstreamCompiler> downstreamCompiler;
+    RefPtr<DownstreamCompiler_Dep1> downstreamCompilerDep1;
 
-    SLANG_RETURN_ON_FAIL(fn(downstreamCompiler));
+    SLANG_RETURN_ON_FAIL(fn(downstreamCompilerDep1));
+
+    ComPtr<IDownstreamCompiler> downstreamCompiler(new DownstreamCompilerAdapter_Dep1(downstreamCompilerDep1));
 
     set->addSharedLibrary(library);
     set->addCompiler(downstreamCompiler);
