+++ source/slang/slang-artifact-output-util.cpp
@@ -47,7 +47,7 @@ static CodeGenTarget _getCodeGenTargetFromDesc(const ArtifactDesc& desc)
     auto downstreamCompiler = getDownstreamCompilerRequiredForTarget(target);
 
     // Get the required downstream compiler
-    DownstreamCompiler* compiler = session->getOrLoadDownstreamCompiler(downstreamCompiler, sink);
+    IDownstreamCompiler* compiler = session->getOrLoadDownstreamCompiler(downstreamCompiler, sink);
 
     if (!compiler)
     {
